K8S-2288: Check Pod Uninitialized Before Kill
Server refuses to work properly until it's had a few calls made to it,
therefore we need to maintain state telling us whether the pod is
initialized (should work) or not (will reject API requests).  In not
initialized, then we can safely delete it, knowing that it's broken as
we have no idea where to restart the initialization process from.  Old
2.0 and 2.1 resources are ignored as we know they can't have the
annotation (at least until topology is updated to upgrade the pod).

First problem is we inject this annotation outside of the usual pod
creation path, so it needs to be preserved as a "3rd party" annotation
(handled by Pat).  Second is we were totally ignoring this annotation
when blowing away pods.

Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/156746
Tested-by: Build Bot <build@couchbase.com>

+++ pkg/cluster/cluster.go
@@ -485,6 +485,11 @@ func (c *Cluster) runReconcile() {
 				continue
 			}
 
+			// Pod is initialized, let the normal reconcile process occur.
+			if _, ok := pod.Annotations[constants.PodInitializedAnnotation]; ok {
+				continue
+			}
+
 			log.Info("Killing uninitialized pod", "cluster", c.namespacedName(), "pod", pod.Name)
 
 			if err := k8sutil.DeletePod(c.k8s, c.cluster.Namespace, pod.Name, metav1.DeleteOptions{}); err != nil {
@@ -680,6 +685,12 @@ func (c *Cluster) recreatePod(m couchbaseutil.Member) error {
 	// Don't delete the volumes here, we need them to recover from, and they
 	// contain precious customer data.
 	if err := c.createPod(ctx, m, *config, false); err != nil {
+		return err
+	}
+
+	// To get here the pod would need to be initialized and clustered, so this is
+	// safe.
+	if err := k8sutil.SetPodInitialized(c.k8s, m.Name()); err != nil {
 		return err
 	}
